pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-rtl_debugger"
version = "0.1.0"
description = "Rtl Debugger environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "cocotb>=2.0.1",
    # Core OpenEnv runtime (provides FastAPI server + HTTP client types)
    # install from github
    # "openenv-core[core] @ git+https://github.com/meta-pytorch/OpenEnv.git",
    # Environment-specific dependencies
    # Add all dependencies needed for your environment here
    # Examples:
    # "numpy>=1.19.0",
    # "torch>=2.0.0",
    # "gymnasium>=0.29.0",
    # "openspiel>=1.0.0",
    # "smolagents>=1.22.0,<2",
    "fastapi>=0.135.2",
    "openai>=2.30.0",
    "openenv-core[core]>=0.2.3",
    "pydantic>=2.12.5",
    "uvicorn>=0.42.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m rtl_debugger.server.app
server = "rtl_debugger.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["rtl_debugger", "rtl_debugger.server"]
package-dir = { "rtl_debugger" = ".", "rtl_debugger.server" = "server" }
